Fix type conversion bug (np.float64 to float)

+++ timbre/timbre.py
@@ -374,7 +374,7 @@ def calc_binary_schedule(datesecs, state1, state2, t_dwell1, t_dwell2, msid, mod
         - Keys in state1 must match Xija component names (e.g. 'pitch', 'ccd_count', 'sim_z')
     """
 
-    num = np.int(duration / (t_dwell1 + t_dwell2))
+    num = int(duration / (t_dwell1 + t_dwell2))
     reltimes = np.cumsum([1, t_dwell1 - 1, 1, t_dwell2 - 1] * num)
     times = np.array(reltimes) - reltimes[0] + datesecs - t_backoff
 
@@ -510,7 +510,7 @@ def find_second_dwell(date, dwell1_state, dwell2_state, t_dwell1, msid, limit, m
 
     # First just check the bounds to avoid unnecessary runs of `opt_fun`
     output = np.array([opt_fun(t) for t in [min_dwell, max_dwell]],
-                      dtype=[('duration2', float64), ('max', float64), ('mean', float64), ('min', float64)])
+                      dtype=[('duration2', float), ('max', float), ('mean', float), ('min', float)])
 
     if 'max' in limit_type:
 
@@ -648,8 +648,8 @@ def _refine_dwell2_time(limit_type, n_dwells, min_dwell, max_dwell, limit, opt_f
                    (max_dwell - min_dwell) * (dwell2_range - dwell2_range[0]) / (dwell2_range[-1] - dwell2_range[0])
 
     # Run the dwell1_state-dwell2_state schedule using the possible dwell 2 guesses
-    output = np.array([opt_fun(t) for t in dwell2_range], dtype=[('duration2', float64), ('max', float64),
-                                                                 ('mean', float64), ('min', float64)])
+    output = np.array([opt_fun(t) for t in dwell2_range], dtype=[('duration2', float), ('max', float),
+                                                                 ('mean', float), ('min', float)])
 
     # Ensure the results are sorted. Although dwell2_range will be sorted, the output may not when two or more dwell
     # times are close, where temperature oscillations from instabilities in the Xija model can cause the results to lose
@@ -672,8 +672,8 @@ def _refine_dwell2_time(limit_type, n_dwells, min_dwell, max_dwell, limit, opt_f
         t_bound = (output_sorted['duration2'][ind - 1], output_sorted['duration2'][ind])
         dwell2_range = np.linspace(np.min(t_bound), np.max(t_bound), n_dwells, endpoint=True)
         output = np.array([opt_fun(t) for t in dwell2_range],
-                          dtype=[('duration2', float64), ('max', float64), ('mean', float64),
-                                 ('min', float64)])
+                          dtype=[('duration2', float), ('max', float), ('mean', float),
+                                 ('min', float)])
 
         # In rare conditions where all 'x' values are very close and 'wobble' a bit, it may not be sorted. If it
         # is not sorted, the quadratic method will result in an error. The linear method is more tolerant of this
@@ -755,7 +755,7 @@ def run_state_pairs(msid, model_spec, init, limit, date, dwell_1_duration, state
                        ({'pitch': 75}, {'pitch': 130}),
                        ({'pitch': 170}, {'pitch': 90}),
                        ({'pitch': 90}, {'pitch': 170}))
-        state_pair_dtype = {'pitch', float64}
+        state_pair_dtype = {'pitch', float}
 
         results = run_state_pairs(msid, model_specs[msid], model_init[msid], limit, date, t_dwell1, state_pairs,
             state_pair_dtype)
@@ -775,16 +775,16 @@ def run_state_pairs(msid, model_spec, init, limit, date, dwell_1_duration, state
 
     base_dtype = [('msid', 'U20'),
                   ('date', 'U8'),
-                  ('datesecs', float64),
-                  ('limit', float64),
-                  ('t_dwell1', float64),
-                  ('t_dwell2', float64),
-                  ('min_temp', float64),
-                  ('mean_temp', float64),
-                  ('max_temp', float64),
-                  ('min_pseudo', float64),
-                  ('mean_pseudo', float64),
-                  ('max_pseudo', float64),
+                  ('datesecs', float),
+                  ('limit', float),
+                  ('t_dwell1', float),
+                  ('t_dwell2', float),
+                  ('min_temp', float),
+                  ('mean_temp', float),
+                  ('max_temp', float),
+                  ('min_pseudo', float),
+                  ('mean_pseudo', float),
+                  ('max_pseudo', float),
                   ('converged', bool),
                   ('unconverged_hot', bool),
                   ('unconverged_cold', bool),
